refactor: remove re-`export`s `index.ts`
