--- src/ViewItems/ViewCreators/index.ts
@@ -1,10 +1,0 @@
-// Copyright (c) Consensys Software Inc. All rights reserved.
-// Licensed under the MIT license.
-
-export * from './GroupViewCreator';
-export * from './NetworkNodeViewCreator';
-export * from './NullableViewCreator';
-export * from './ProjectViewCreator';
-export * from './ServiceViewCreator';
-export * from './ViewCreator';
-export * from './LayerViewCreator';
