+++ src/commands/InfuraCommands.ts
@@ -3,7 +3,7 @@
 
 import {Constants} from '../Constants';
 import {showQuickPickMany} from '../helpers/userInteraction';
-import {InfuraResourceExplorer} from '../resourceExplorers';
+import {InfuraResourceExplorer} from '@/resourceExplorers/InfuraResourceExplorer';
 import {InfuraServiceClient} from '@/services/infuraService/InfuraServiceClient';
 
 export namespace InfuraCommands {
