+++ src/commands/ServiceCommands.ts
@@ -13,7 +13,9 @@ import {InfuraService} from '@/Models/TreeItems/InfuraService';
 import {LocalService} from '@/Models/TreeItems/LocalService';
 import {Project} from '@/Models/TreeItems/Project';
 import {Service, ServiceTypes} from '@/Models/TreeItems/Service';
-import {GenericResourceExplorer, InfuraResourceExplorer, LocalResourceExplorer} from '@/resourceExplorers';
+import {GenericResourceExplorer} from '@/resourceExplorers/GenericResourceExplorer';
+import {InfuraResourceExplorer} from '@/resourceExplorers/InfuraResourceExplorer';
+import {LocalResourceExplorer} from '@/resourceExplorers/LocalResourceExplorer';
 import {GanacheService} from '@/services/ganache/GanacheService';
 import {TreeManager} from '@/services/tree/TreeManager';
 import {Telemetry} from '@/TelemetryClient';
