--- src/commands/index.ts
@@ -1,12 +1,0 @@
-// Copyright (c) Consensys Software Inc. All rights reserved.
-// Licensed under the MIT license.
-
-export * from './ContractCommands';
-export * from './GanacheCommands';
-export * from './InfuraCommands';
-export * from './ProjectCommands';
-export * from './ServiceCommands';
-export * from './TruffleCommands';
-export * from './SdkCoreCommands';
-export * from './GenericCommands';
-export * from './DashboardCommands';
