--- src/helpers/index.ts
@@ -1,10 +1,0 @@
-// Copyright (c) Consensys Software Inc. All rights reserved.
-// Licensed under the MIT license.
-
-import * as outputCommandHelper from './command';
-import {extractEnumsInfo, extractEnumsInfoSafe} from './enumExtractor';
-import * as shell from './shell';
-
-const spawnProcess = outputCommandHelper.spawnProcess;
-
-export {extractEnumsInfo, extractEnumsInfoSafe, outputCommandHelper, shell, spawnProcess};
