--- src/resourceExplorers/index.ts
@@ -1,7 +1,0 @@
-// Copyright (c) Consensys Software Inc. All rights reserved.
-// Licensed under the MIT license.
-
-export * from './InfuraResourceExplorer';
-export * from './LocalResourceExplorer';
-export * from './InfuraResourceExplorer';
-export * from './GenericResourceExplorer';
