+++ test/commands/DashboardCommands.test.ts
@@ -5,7 +5,7 @@ import assert from 'assert';
 import {ChildProcess} from 'child_process';
 import sinon from 'sinon';
 import {OutputChannel, window} from 'vscode';
-import {DashboardCommands} from '@/commands';
+import {DashboardCommands} from '@/commands/DashboardCommands';
 import {Constants, RequiredApps} from '@/Constants';
 import {required} from '@/helpers/required';
 import * as shell from '../../src/helpers/shell';
