+++ test/commands/DashboardService.test.ts
@@ -7,7 +7,7 @@ import sinon from 'sinon';
 import {Constants, RequiredApps} from '../../src/Constants';
 import stream from 'stream';
 import {OutputChannel, window} from 'vscode';
-import * as outputCommandHelper from '../../src/helpers';
+import * as outputCommandHelper from '@/helpers/command';
 import * as shell from '../../src/helpers/shell';
 import {DashboardService} from '@/services/dashboard/DashboardService';
 import * as DashboardServiceClient from '../../src/services/dashboard/DashboardServiceClient';
